+++ server/models/units.js
@@ -10,4 +10,5 @@ module.exports = function( sequelize, DataTypes ){
   });
 
   return Unit;
+     models.Unit.hasMany( models.Ingredient );
 }
